refactor: split the public API headers into constants + data types
Extract three headers under library/src/api/ so internal solver code no
longer includes the external function-declaration surface:

- dds_constants.hpp: bridge dimensions (moved from utility/constants.h),
  DDS_VERSION / MAXNOOFBOARDS / MAXNOOFTABLES, the RETURN_* / TEXT_* status
  codes, solver tuning constants, HAND_ID, and the DLLEXPORT / STDCALL /
  EXTERN_C macros. Dependency-free (//library/src/api:dds_constants).
- dds_c_data_types.h: the six POD structs that cross the pure-C shim
  (Deal, FutureTricks, DdTableDeal, DdTableDealPBN, DdTableResults,
  ParResults). Included by dds_data_types.hpp.
- dds_data_types.hpp: every other legacy struct plus the internal search
  structures.

dll.h now carries only the flat function declarations; api/dds.h is a thin
internal aggregator of the two new data/constant headers. Internal headers
and sources include the new headers directly; the flat-API implementation
files (solve_board, solver_if, calc_tables, init, play_analyser, par) now
include <api/dll.h> explicitly. Updates the Visual Studio project file lists
and the dds-public-api / constants-and-debug specs.
