+++ library/src/api/calc_par.hpp
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include <api/dds.h>
+#include <api/dds_data_types.hpp>
 #include <solver_context/solver_context.hpp>
 
 // Naming note: New C++ APIs in DDS 3 use snake_case (calc_par, calc_par_from_table).
