+++ library/src/api/dll.h
@@ -10,447 +10,18 @@
 
 #pragma once
 
-#include <utility/constants.h>
-
-#if (defined(_WIN32) || defined(__CYGWIN__)) && ! defined(__clang__)
-  #define DLLEXPORT __declspec(dllexport)
-  #define STDCALL __stdcall
-#else
-  #define DLLEXPORT
-  #define STDCALL
-#endif
-
-#ifdef __cplusplus
-  #define EXTERN_C extern "C"
-#else
-  #define EXTERN_C
-  #include <stdbool.h> // make "bool" available
-#endif
-
-/* Version 3.1.0. Allowing for 2 digit minor versions */
-// These three stay object-like macros: this is the frozen legacy C API
-// header and external consumers conventionally test the version / limits in
-// the preprocessor (e.g. #if DDS_VERSION >= 30100, #ifdef MAXNOOFBOARDS).
-#define DDS_VERSION 30100
-
-#define MAXNOOFBOARDS 200
-
-#define MAXNOOFTABLES 40
-
-
-// Error codes. See interface document for more detail.
-// Call ErrorMessage(code, line[]) to get the text form in line[].
-
-// Success.
-constexpr int RETURN_NO_FAULT = 1;
-constexpr const char TEXT_NO_FAULT[] = "Success";
-
-// Currently happens when fopen() fails or when AnalyseAllPlaysBin()
-// get a different number of Boards in its first two arguments.
-constexpr int RETURN_UNKNOWN_FAULT = -1;
-constexpr const char TEXT_UNKNOWN_FAULT[] = "General error";
-
-// SolveBoard()
-constexpr int RETURN_ZERO_CARDS = -2;
-constexpr const char TEXT_ZERO_CARDS[] = "Zero cards";
-
-// SolveBoard()
-constexpr int RETURN_TARGET_TOO_HIGH = -3;
-constexpr const char TEXT_TARGET_TOO_HIGH[] =
-  "Target exceeds number of tricks";
-
-// SolveBoard()
-constexpr int RETURN_DUPLICATE_CARDS = -4;
-constexpr const char TEXT_DUPLICATE_CARDS[] = "Cards duplicated";
-
-// SolveBoard()
-constexpr int RETURN_TARGET_WRONG_LO = -5;
-constexpr const char TEXT_TARGET_WRONG_LO[] =
-  "Target is less than -1";
-
-// SolveBoard()
-constexpr int RETURN_TARGET_WRONG_HI = -7;
-constexpr const char TEXT_TARGET_WRONG_HI[] =
-  "Target is higher than 13";
-
-// SolveBoard()
-constexpr int RETURN_SOLNS_WRONG_LO = -8;
-constexpr const char TEXT_SOLNS_WRONG_LO[] =
-  "Solutions parameter is less than 1";
-
-// SolveBoard()
-constexpr int RETURN_SOLNS_WRONG_HI = -9;
-constexpr const char TEXT_SOLNS_WRONG_HI[] =
-  "Solutions parameter is higher than 3";
-
-// SolveBoard(), self-explanatory.
-constexpr int RETURN_TOO_MANY_CARDS = -10;
-constexpr const char TEXT_TOO_MANY_CARDS[] = "Too many cards";
-
-// SolveBoard()
-constexpr int RETURN_SUIT_OR_RANK = -12;
-constexpr const char TEXT_SUIT_OR_RANK[] =
-  "currentTrickSuit or currentTrickRank has wrong data";
-
-// SolveBoard
-constexpr int RETURN_PLAYED_CARD = -13;
-constexpr const char TEXT_PLAYED_CARD[] =
-  "Played card also remains in a hand";
-
-// SolveBoard()
-constexpr int RETURN_CARD_COUNT = -14;
-constexpr const char TEXT_CARD_COUNT[] =
-  "Wrong number of remaining cards in a hand";
-
-// SolveBoard()
-constexpr int RETURN_THREAD_INDEX = -15;
-constexpr const char TEXT_THREAD_INDEX[] =
-  "Thread index is not 0 .. maximum";
-
-// SolveBoard()
-constexpr int RETURN_MODE_WRONG_LO = -16;
-constexpr const char TEXT_MODE_WRONG_LO[] =
-  "Mode parameter is less than 0";
-
-// SolveBoard()
-constexpr int RETURN_MODE_WRONG_HI = -17;
-constexpr const char TEXT_MODE_WRONG_HI[] =
-  "Mode parameter is higher than 2";
-
-// SolveBoard()
-constexpr int RETURN_TRUMP_WRONG = -18;
-constexpr const char TEXT_TRUMP_WRONG[] = "Trump is not in 0 .. 4";
-
-// SolveBoard()
-constexpr int RETURN_FIRST_WRONG = -19;
-constexpr const char TEXT_FIRST_WRONG[] = "First is not in 0 .. 2";
-
-// AnalysePlay*() family of functions.
-// (a) Less than 0 or more than 52 cards supplied.
-// (b) Invalid suit or rank supplied.
-// (c) A played card is not held by the right player.
-constexpr int RETURN_PLAY_FAULT = -98;
-constexpr const char TEXT_PLAY_FAULT[] = "AnalysePlay input error";
-
-// Returned from a number of places if a PBN string is faulty.
-constexpr int RETURN_PBN_FAULT = -99;
-constexpr const char TEXT_PBN_FAULT[] = "PBN string error";
-
-// SolveBoard() and AnalysePlay*()
-constexpr int RETURN_TOO_MANY_BOARDS = -101;
-constexpr const char TEXT_TOO_MANY_BOARDS[] =
-  "Too many Boards requested";
-
-// Returned from multi-threading functions.
-constexpr int RETURN_THREAD_CREATE = -102;
-constexpr const char TEXT_THREAD_CREATE[] =
-  "Could not create threads";
-
-// Returned from multi-threading functions when something went
-// wrong while waiting for all threads to complete.
-constexpr int RETURN_THREAD_WAIT = -103;
-constexpr const char TEXT_THREAD_WAIT[] =
-  "Something failed waiting for thread to end";
-
-// Tried to set a multi-threading system that is not present in DLL.
-constexpr int RETURN_THREAD_MISSING = -104;
-constexpr const char TEXT_THREAD_MISSING[] =
-  "Multi-threading system not present";
-
-// CalcAllTables*()
-constexpr int RETURN_NO_SUIT = -201;
-constexpr const char TEXT_NO_SUIT[] =
-  "Denomination filter vector has no entries";
-
-// CalcAllTables*()
-constexpr int RETURN_TOO_MANY_TABLES = -202;
-constexpr const char TEXT_TOO_MANY_TABLES[] =
-  "Too many DD tables requested";
-
-// SolveAllChunks*()
-constexpr int RETURN_CHUNK_SIZE = -301;
-constexpr const char TEXT_CHUNK_SIZE[] = "Chunk size is less than 1";
-
-// Par(), SidesPar(), SidesParBin(), DealerPar(), DealerParBin()
-constexpr int RETURN_PAR_TABLE_FAULT = -401;
-constexpr const char TEXT_PAR_TABLE_FAULT[] =
-  "Missing double dummy table, or an entry outside the range 0 to 13";
-
-
-
-/**
- * @brief Stores the result of a double dummy analysis for a single position.
- *
- * Contains the number of nodes searched, the number of cards in the result,
- * and arrays for each card's suit, rank, equality group, and score.
- */
-struct FutureTricks
-{
-  int nodes;
-  int cards;
-  int suit[13];
-  int rank[13];
-  int equals[13];
-  int score[13];
-};
-
-/**
- * @brief Represents a bridge Deal for double dummy analysis.
- *
- * @param trump The trump suit (0 = NT, 1 = Spades, ...)
- * @param first The hand to play first (0 = N, 1 = E, ...)
- * @param currentTrickSuit Suits of cards played in the current trick
- * @param currentTrickRank Ranks of cards played in the current trick
- * @param remainCards Remaining cards in each hand and suit
- */
-struct Deal
-{
-  int trump;
-  int first;
-  int currentTrickSuit[3];
-  int currentTrickRank[3];
-  unsigned int remainCards[DDS_HANDS][DDS_SUITS];
-};
-
-
-/**
- * @brief Represents a bridge Deal in PBN (Portable Bridge Notation) format.
- *
- * @param trump The trump suit
- * @param first The hand to play first
- * @param currentTrickSuit Suits of cards played in the current trick
- * @param currentTrickRank Ranks of cards played in the current trick
- * @param remainCards PBN string describing remaining cards. Only the first hand may have a compass letter (N/E/S/W); later hands follow clockwise with no extra directions.
- */
-struct DealPBN
-{
-  int trump;
-  int first;
-  int currentTrickSuit[3];
-  int currentTrickRank[3];
-  char remainCards[80];
-};
-
-
-/**
- * @brief Represents multiple bridge deals for batch analysis.
- *
- * @param noOfBoards Number of deals
- * @param deals Array of deals
- * @param target Array of targets for each Deal
- * @param solutions Array of solution modes for each Deal
- * @param mode Array of modes for each Deal
- */
-struct Boards
-{
-  int no_of_boards;
-  struct Deal deals[MAXNOOFBOARDS];
-  int target[MAXNOOFBOARDS];
-  int solutions[MAXNOOFBOARDS];
-  int mode[MAXNOOFBOARDS];
-};
-
-/**
- * @brief Multiple boards in PBN format for batch solving.
- *
- * Similar to Boards but uses PBN (Portable Bridge Notation) format
- * for deal representation. Used for solving multiple boards efficiently.
- *
- * @see Boards
- */
-struct BoardsPBN
-{
-  int no_of_boards;                            ///< Number of boards to solve
-  struct DealPBN deals[MAXNOOFBOARDS];       ///< Array of deals in PBN format
-  int target[MAXNOOFBOARDS];                 ///< Target tricks for each board
-  int solutions[MAXNOOFBOARDS];              ///< Solution mode for each board
-  int mode[MAXNOOFBOARDS];                   ///< Solve mode for each board
-};
-
-/**
- * @brief Solutions for multiple boards.
- *
- * Container for results from batch board solving operations.
- * Each entry contains the complete future tricks analysis for one board.
- *
- * @see FutureTricks
- */
-struct SolvedBoards
-{
-  int no_of_boards;                                    ///< Number of solved boards
-  struct FutureTricks solved_board[MAXNOOFBOARDS];    ///< Array of solutions
-};
-
-struct DdTableDeal
-{
-  unsigned int cards[DDS_HANDS][DDS_SUITS];
-};
-
-struct DdTableDeals
-{
-  int no_of_tables;
-  struct DdTableDeal deals[MAXNOOFTABLES * DDS_STRAINS];
-};
-
-struct DdTableDealPBN
-{
-  char cards[80];
-};
-
-struct DdTableDealsPBN
-{
-  int no_of_tables;
-  struct DdTableDealPBN deals[MAXNOOFTABLES * DDS_STRAINS];
-};
-
-struct DdTableResults
-{
-  int res_table[DDS_STRAINS][DDS_HANDS];
-};
-
-struct DdTablesRes
-{
-  int no_of_boards;
-  struct DdTableResults results[MAXNOOFTABLES * DDS_STRAINS];
-};
-
-struct ParResults
-{
-  /* index = 0 is NS view and index = 1
-     is EW view. By 'view' is here meant
-     which side that starts the bidding. */
-  char par_score[2][16];
-  char par_contracts_string[2][128];
-};
-
-struct AllParResults
-{
-  struct ParResults par_results[MAXNOOFTABLES];
-};
-
-struct ParResultsDealer
-{
-  /* number: Number of contracts yielding the par score.
-     score: Par score for the specified dealer hand.
-     contracts:  Par contract text strings.  The first contract
-       is in contracts[0], the last one in contracts[number-1].
-       The detailed text format is is given in the DLL interface
-       document.
-  */
-  int number;
-  int score;
-  char contracts[10][10];
-};
-
-struct ContractType
-{
-  int under_tricks; /* 0 = make 1-13 = sacrifice */
-  int over_tricks; /* 0-3, e.g. 1 for 4S + 1. */
-  int level; /* 1-7 */
-  int denom; /* 0 = No Trumps, 1 = trump Spades, 2 = trump Hearts,
-                  3 = trump Diamonds, 4 = trump Clubs */
-  int seats; /* One of the cases N, E, W, S, NS, EW;
-                   0 = N 1 = E, 2 = S, 3 = W, 4 = NS, 5 = EW */
-};
-
-struct ParResultsMaster
-{
-  int score; /* Sign according to the NS view */
-  int number; /* Number of contracts giving the par score */
-  struct ContractType contracts[10]; /* Par contracts */
-};
-
-struct ParTextResults
-{
-  char par_text[2][128]; /* Short text for par information, e.g.
-            Par -110: EW 2S EW 2D+1 */
-  bool equal; /* true in the normal case when it does not matter who
-            starts the bidding. Otherwise, false. */
-};
-
-
-struct PlayTraceBin
-{
-  int number;
-  int suit[52];
-  int rank[52];
-};
-
-struct PlayTracePBN
-{
-  int number;
-  char cards[106];
-};
-
-struct SolvedPlay
-{
-  int number;
-  int tricks[53];
-};
-
-struct PlayTracesBin
-{
-  int no_of_boards;
-  struct PlayTraceBin plays[MAXNOOFBOARDS];
-};
-
-struct PlayTracesPBN
-{
-  int no_of_boards;
-  struct PlayTracePBN plays[MAXNOOFBOARDS];
-};
-
-struct SolvedPlays
-{
-  int no_of_boards;
-  struct SolvedPlay solved[MAXNOOFBOARDS];
-};
-
-struct DDSInfo
-{
-  // Version 2.8.0 has 2, 8, 0 and a string of 2.8.0
-  int major, minor, patch; 
-  char version_string[10];
-
-  // Currently 0 = unknown, 1 = Windows, 2 = Cygwin, 3 = Linux, 4 = Apple
-  int system;
-
-  // We know 32 and 64-bit systems.
-  int numBits;
-
-  // Currently 0 = unknown, 1 = Microsoft Visual C++, 2 = mingw,
-  // 3 = GNU g++, 4 = clang
-  int compiler;
-
-  // Currently 0 = none, 1 = DllMain, 2 = Unix-style
-  int constructor;
-
-  int numCores;
-
-  // Currently 
-  // 0 = none, 
-  // 1 = Windows (native), 
-  // 2 = OpenMP, 
-  // 3 = GCD,
-  // 4 = Boost,
-  // 5 = STL,
-  // 6 = TBB,
-  // 7 = STLIMPL (for_each), experimental only
-  // 8 = PPLIMPL (for_each), experimental only
-  int threading;
-
-  // The actual number of threads configured
-  int noOfThreads;
-
-  // This will break if there are > 128 threads...
-  // The string is of the form LLLSSS meaning 3 large TT memories
-  // and 3 small ones.
-  char threadSizes[128];
-
-  char systemString[1024];
-};
-
-
+/// @file dll.h
+/// @brief The flat legacy C API: the historical Haglund/Hein entry points.
+///
+/// This is the public function-declaration surface. It is included only by API
+/// consumers (examples, bindings, tests) and by the API implementation files
+/// that define these symbols — never by internal solver code, which takes the
+/// data model from <api/dds_data_types.hpp> and the constants from
+/// <api/dds_constants.hpp> directly.
+
+#include <api/dds_constants.hpp>   // DLLEXPORT / STDCALL / EXTERN_C, DDS_VERSION,
+                                   // MAXNOOFBOARDS, MAXNOOFTABLES, RETURN_* / TEXT_*
+#include <api/dds_data_types.hpp>  // struct Deal, Boards, FutureTricks, DdTable*, Par*, ...
 
 /**
  * @brief Initialize the solver's static memory.
@@ -798,7 +369,7 @@ EXTERN_C DLLEXPORT auto STDCALL DealerPar(
 EXTERN_C DLLEXPORT auto STDCALL DealerParBin(
   struct DdTableResults const * tablep,
   struct ParResultsMaster * presp,
-  int dealer, 
+  int dealer,
   int vulnerable) -> int;
 
 EXTERN_C DLLEXPORT auto STDCALL SidesParBin(
