+++ library/src/calc_tables.hpp
@@ -11,7 +11,7 @@
 
 #include <vector>
 
-#include <api/dll.h>
+#include <api/dds_data_types.hpp>
 #include <solver_context/solver_context.hpp>
