+++ library/src/dump.hpp
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include <api/dds.h>
+#include <api/dds_data_types.hpp>
 #include <moves/moves.hpp>
 #include <solver_context/solver_context.hpp>
