+++ library/src/heuristic_sorting/heuristic_sorting.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <api/dds.h>
+#include <api/dds_data_types.hpp>
 
 /// @brief Track information for maintaining position state during move generation.
 ///
