+++ library/src/init.hpp
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include <api/dds.h>
+#include <api/dds_data_types.hpp>
 #include <system/memory.hpp>
 #include <solver_context/solver_context.hpp>
 #include <memory>
