+++ library/src/play_analyser.hpp
@@ -11,7 +11,7 @@
 
 #include <vector>
 
-#include <api/dll.h>
+#include <api/dds_data_types.hpp>
 
 
 void detect_play_duplicates(
