+++ library/src/system/deal_fanout.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <api/dll.h>
+#include <api/dds_data_types.hpp>
 
 namespace dds
 {
