+++ library/src/system/memory.hpp
@@ -12,7 +12,7 @@
 
 #include <vector>
 
-#include <api/dds.h>
+#include <api/dds_data_types.hpp>
 #include <moves/moves.hpp>
 #include <system/thread_data.hpp>
