+++ library/src/system/parallel_boards.cpp
@@ -21,10 +21,7 @@
 #include <utility>
 #include <vector>
 
-#if defined(__EMSCRIPTEN__)
-#include <api/dds.h>
-#endif
-#include <api/dll.h>
+#include <api/dds_data_types.hpp>
 
 
 namespace
