+++ library/src/utility/BUILD.bazel
@@ -16,6 +16,6 @@ cc_library(
          "//library/src:__subpackages__",
          "//library/tests:__subpackages__"
     ],
-    deps = [],
+    deps = ["//library/src/api:dds_constants"],
 )
