+++ library/tests/solve_board/analyse_play_consistency.cpp
@@ -21,6 +21,7 @@
 
 // Project headers
 #include <api/dds.h>
+#include <api/dll.h>
 
 namespace {
